refactor: add form card component
