+++ src/components/CardForm/Props.tsx
@@ -1,0 +1,6 @@
+import { ICard } from 'types';
+
+export type Props = {
+  cardList: ICard[];
+  setCardList: React.Dispatch<React.SetStateAction<ICard[]>>;
+};
